rename var to variable in file paths
