+++ src/api/variables/mutate/deleteVariable.ts
