+++ src/api/variables/schema.ts
