+++ src/pages/namespace/Settings/variables/VariablesList.tsx
@@ -5,14 +5,14 @@ import { Table, TableBody } from "~/design/Table";
 
 import Button from "~/design/Button";
 import { Card } from "~/design/Card";
-import Create from "./variables/Create";
-import Delete from "./variables/Delete";
-import Edit from "./variables/Edit";
-import ItemRow from "./ItemRow";
-import { VarSchemaType } from "~/api/vars/schema";
-import { useDeleteVar } from "~/api/vars/mutate/deleteVar";
+import Create from "./Create";
+import Delete from "./Delete";
+import Edit from "./Edit";
+import ItemRow from "../ItemRow";
+import { VarSchemaType } from "~/api/variables/schema";
+import { useDeleteVar } from "~/api/variables/mutate/deleteVariable";
 import { useTranslation } from "react-i18next";
-import { useVars } from "~/api/vars/query/useVars";
+import { useVars } from "~/api/variables/query/useVariables";
 
 const VariablesList: FC = () => {
   const { t } = useTranslation();
